fixed faulty termination condition
previously the algorithm would terminate as soon as the end node is in
the neighbor list. it did not check, whether the weight is actually
better or worse in that case. this has been fixed, however there is
still a faulty behaviour: if two edges share the same weight, the
algorithm does not necessarily pick the correct shortest path. the
current console.log proves this.